requirements.txt
paddlepaddle-gpu >= 2.2.0
paddleslim >= 2.2.1
visualdl >= 2.1.1
opencv-contrib-python == 4.3.0.38
numba == 0.53.1
scikit-learn == 0.23.2
scikit-image >= 0.14.0
pandas
scipy
cython
pycocotools
shapely
lap
motmetrics
chardet
openpyxl
easydict
munch
natsort
geojson

# # Self installation
# GDAL >= 3.1.3
